refactor: rename file, extract out refCallback
